+++ platform/LocalFileSystem.h
@@ -51,7 +51,7 @@ public:
 
     virtual int sync();
 
-    virtual size_t size();
+    virtual off_t size();
 
 protected:
     virtual void lock();
